Changed to router to allow for hat direct links
