+++ package.json
@@ -11,6 +11,7 @@
     "@popperjs/core": "^2.11.4",
     "bootstrap": "^5.1.3",
     "vue": "^3.2.31",
+    "vue-router": "^4.0.14",
     "vuejs-paginate-next": "^1.0.2"
   },
   "devDependencies": {
